Day4-Django-advanced/django-starter/stock/serializers/stock_serializer.py
from rest_framework import serializers
from stock.models.stock import Stock

class StockSerializer(serializers.ModelSerializer):
    class Meta:
        model = Stock
        fields = ['id', 'company_name', 'current_price']