Day4-Django-advanced/pyproject.toml
[project]
name = "day4-django-advanced"
version = "0.1.0"
description = ""
authors = [
    {name = "ranjanlamsal",email = "lamsalranjan19@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pytz (>=2025.1,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
